Change receptive field training samples from a integer to a float fraction of input area, so it needs less tuning.

+++ code/core/src/io/agi/core/alg/RegionFactory.java
@@ -66,7 +66,7 @@ public class RegionFactory {
             int classifierHeightCells,
 
             // Organizer training
-            int receptiveFieldsTrainingSamples,
+            float receptiveFieldsTrainingSamples,
             int receptiveFieldSize,
             float organizerLearningRate,
             float organizerLearningRateNeighbours,
